correcting login and register feedback
