+++ client/storefront/apolo-client.tsx
@@ -5,6 +5,7 @@ import {
   InMemoryCache,
 } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
+import * as ls from 'local-storage';
 // import dotenv from 'dotenv';
 // dotenv.config();
 // const { APOLLO_CLIENT_URL } = process.env;
@@ -16,7 +17,7 @@ const httpLink = createHttpLink({
 const authLink = setContext((_, { headers }) => {
   let token;
   if (typeof window !== 'undefined') {
-    token = localStorage.getItem('token');
+    ls.get('token');
   }
   return {
     headers: { ...headers, Authorization: token },
