+++ client/storefront/package.json
@@ -21,6 +21,7 @@
     "eslint-config-next": "11.0.1",
     "graphql": "^16.6.0",
     "jwt-decode": "^3.1.2",
+    "local-storage": "^2.0.0",
     "next": "13.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0"
