+++ client/storefront/pages/login.tsx
@@ -1,96 +1,7 @@
-import { useMutation } from '@apollo/client';
-import { useRouter } from 'next/router';
-import { useState } from 'react';
-import client from '../apolo-client';
-import { useAuth } from '../src/context/authContext';
-import { LOGIN_USER } from '../src/queries/queries';
-
-function LoginUser() {
-  const { login, user, logout } = useAuth();
-  const router = useRouter();
-
-  const [loginUser, { loading, error }] = useMutation(LOGIN_USER, {
-    client: client,
-    onCompleted: async function (data) {
-      localStorage.setItem('token', data.loginUser.token);
-      login({ token: data.loginUser.token, email: data.loginUser.email });
-      router.push('/CategoriesHome');
-    },
-  });
-
-  const [newUser, setNewUser] = useState({
-    email: '',
-    password: '',
-  });
-
-  const handleChange = (e: any) => {
-    setNewUser({
-      ...newUser,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const handleClick = async (e: any) => {
-    e.preventDefault();
-    try {
-      await loginUser({
-        variables: {
-          user: newUser,
-        },
-      });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
-  if (loading) return <p>Loading...</p>;
-  if (error) return <p>Error : {error.message}</p>;
-
-  return (
-    <form>
-      <p className='mb-4'>Please login to your account</p>
-      <div className='mb-4'>
-        <input
-          type='text'
-          name='email'
-          className='form-control block w-full px-3 py-1.5 text-base font-normal text-gray-700 bg-white bg-clip-padding border border-solid border-gray-300 rounded transition ease-in-out m-0 focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none'
-          id='exampleFormControlInput1'
-          placeholder='E-mail'
-          onChange={(e) => handleChange(e)}
-        />
-      </div>
-      <div className='mb-4'>
-        <input
-          type='password'
-          name='password'
-          className='form-control block w-full px-3 py-1.5 text-base font-normal text-gray-700 bg-white bg-clip-padding border border-solid border-gray-300 rounded transition ease-in-out m-0 focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none'
-          id='exampleFormControlInput1'
-          placeholder='Password'
-          onChange={(e) => handleChange(e)}
-        />
-      </div>
-      <div className='text-center pt-1 mb-12 pb-1'>
-        <button
-          onClick={(e) => handleClick(e)}
-          className='bg-teal-700 inline-block px-6 py-2.5 text-white font-medium text-xs leading-tight uppercase rounded shadow-md hover:bg-teal-500 hover:shadow-lg focus:shadow-lg focus:outline-none focus:ring-0 active:shadow-lg transition duration-150 ease-in-out w-full mb-3'
-          type='button'
-          data-mdb-ripple='true'
-          data-mdb-ripple-color='light'
-        >
-          Log in
-        </button>
-      </div>
-    </form>
-  );
-}
+import Link from 'next/link';
+import LoginUser from '../src/components/LoginUser';
 
 export default function login() {
-  const router = useRouter();
-
-  function redirect() {
-    router.push('/register');
-  }
-
   return (
     <section className='h-full gradient-form bg-gray-900 md:h-screen'>
       <div className='container py-12 px-6 h-full'>
@@ -113,15 +24,15 @@ export default function login() {
                     <LoginUser />
                     <div className='flex items-center justify-between pb-6'>
                       <p className='mb-0 mr-2'>Don't have an account?</p>
-                      <button
-                        onClick={redirect}
+                      <Link
+                        href='/register'
                         type='button'
                         className='inline-block px-6 py-2 border-2 border-teal-800 text-teal-800 font-medium text-xs leading-tight uppercase rounded hover:bg-black hover:bg-opacity-5 focus:outline-none focus:ring-0 transition duration-150 ease-in-out'
                         data-mdb-ripple='true'
                         data-mdb-ripple-color='light'
                       >
                         Sign up
-                      </button>
+                      </Link>
                     </div>
                   </div>
                 </div>
